// src/sdks/computation/computationV2ApiDocs.ts
export interface ControlPlanData {
  activityDesc?: string;

  calculateDesc?: string;

  calculateType?: string;

  categoryDesc?: string;

  classify?: number;

  collectDesc?: string;

  computationFlag?: boolean;

  controlPlanId?: number;

  createBy?: number;

  createTime?: Date;

  ghgCategory?: '1' | '2' | '3';

  ghgClassify?:
    | '1'
    | '10'
    | '11'
    | '12'
    | '13'
    | '14'
    | '15'
    | '16'
    | '17'
    | '18'
    | '19'
    | '2'
    | '20'
    | '21'
    | '22'
    | '23'
    | '24'
    | '25'
    | '3'
    | '4'
    | '5'
    | '6'
    | '7'
    | '8'
    | '9';

  id?: number;

  isoCategory?: '1' | '2' | '3' | '4' | '5' | '6';

  isoClassify?:
    | '1'
    | '10'
    | '11'
    | '12'
    | '13'
    | '14'
    | '15'
    | '16'
    | '17'
    | '18'
    | '19'
    | '2'
    | '20'
    | '21'
    | '22'
    | '23'
    | '24'
    | '25'
    | '26'
    | '3'
    | '4'
    | '5'
    | '6'
    | '7'
    | '8'
    | '9';

  standardType?: '1' | '2';

  storageDesc?: string;

  updateBy?: number;

  updateTime?: Date;
  languageSourceList?: any[];
}
// ...
export interface ControlPlanOrg {
  controlPlanId?: number;

  createBy?: number;

  createTime?: Date;

  deptName?: string;

  gasGroupDesc?: string;

  gasGroupImg?: string;

  id?: number;

  intro?: string;

  orgId?: number;

  planeImg?: string;

  planeImgDesc?: string;

  produceAddress?: string;

  produceArea?: string;

  produceAreaCode?: number;

  produceCodes?: number[];

  regAddress?: string;

  regArea?: string;

  regAreaCode?: number;

  regCodes?: number[];

  updateBy?: number;

  updateTime?: Date;
  languageSourceList?: any[];
}
// ...
export interface ControlPlanProduct {
  controlPlanId?: number;

  createBy?: number;

  createTime?: Date;

  deleted?: boolean;

  id?: number;

  serviceDesc?: string;

  serviceName?: string;

  serviceUnit?: string;

  updateBy?: number;

  updateTime?: Date;
}
// ...
export interface ControlPlanResp {
  borderChange?: string;

  borderDesc?: string;

  borderMethod?: '1' | '2' | '3';

  companyId?: number;
  createBy?: number;
  createTime?: Date;

  dataQuality?: string;

  ghgList?: ControlPlanData[];

  id?: number;

  isoList?: ControlPlanData[];

  org?: ControlPlanOrg;

  orgId?: number;

  planContent?: string;

  planDate?: string;

  productList?: ControlPlanProduct[];

  remark?: string;
  updateBy?: number;
  updateTime?: Date;

  version?: string;
  languageSourceList?: any[];
}